requirement.txt
aspy.yaml==1.3.0
authy==2.2.3
certifi==2019.3.9
cfgv==1.6.0
chardet==3.0.4
Click==7.0
Django==2.2.1
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
docopt==0.4.0
filelock==3.0.12
geographiclib==1.49
geopy==1.20.0
identify==1.4.3
idna==2.8
importlib-metadata==0.15
importlib-resources==1.0.2
lxml==4.3.3
mailchimp==2.0.9
mock==3.0.5
mysqlclient==1.4.2.post1
nodeenv==1.3.3
phonenumbers==8.8.4
Pillow==6.0.0
pkg-resources==0.0.0
pluggy==0.11.0
pre-commit==1.16.1
py==1.8.0
PyJWT==1.7.1
python-dotenv==0.7.1
pytz==2019.1
PyYAML==5.1
requests==2.22.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
toml==0.10.0
tox==3.12.1
translate==3.5.0
urllib3==1.25.3
virtualenv==16.6.0
zipp==0.5.1
